+++ YummyTimeWeb/src/main/java/servlets/menuServlets/AboutUsServlet.java
@@ -1,10 +1,12 @@
-package servletsDoGET;
+package servlets.menuServlets;
 
-import dao.*;
+import dao.SessionBean;
+import dao.TemplateProvider;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 import javax.inject.Inject;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -13,19 +15,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-@WebServlet("/statistics")
-public class StatisticsServlet extends HttpServlet {
+@WebServlet("/about")
+public class AboutUsServlet extends HttpServlet {
 
-    private static Logger logger = LoggerFactory.getLogger(ViewRecipeServlet.class);
+    private static Logger logger = LoggerFactory.getLogger(ContactUsServlet.class);
 
     @Inject
     private TemplateProvider templateProvider;
-
-    @Inject
-    private RecipeDao recipeDao;
 
     @Inject
     SessionBean sessionBean;
@@ -33,19 +31,8 @@ public class StatisticsServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        Template template = templateProvider.getTemplate(getServletContext(), "statisticsWeb.ftlh");
+        Template template = templateProvider.getTemplate(getServletContext(), "aboutUs.ftlh");
         Map<String, Object> model = new HashMap<>();
-
-        List<Statistic> categoryStatistic = recipeDao.categoryStatistics();
-        List<Statistic> recipeStatistic = recipeDao.statisticRecipe();
-
-        if (!categoryStatistic.isEmpty()) {
-            model.put("categoryStatistic", categoryStatistic);
-        }
-
-        if (!recipeStatistic.isEmpty()) {
-            model.put("recipeStatistic", recipeStatistic);
-        }
 
         if(sessionBean.getLogged()){
             model.put("logged", "yes");
@@ -55,7 +42,7 @@ public class StatisticsServlet extends HttpServlet {
             template.process(model, resp.getWriter());
         } catch (TemplateException e) {
             e.printStackTrace();
-            logger.warn("View statistics cannot be loaded template!");
+            logger.warn("View about us cannot be loaded template!");
         }
     }
 }
