+++ YummyTimeWeb/src/main/java/servlets/menuServlets/ContactUsServlet.java
@@ -1,11 +1,11 @@
-package servletsDoGET;
+package servlets.menuServlets;
 
+import dao.SessionBean;
 import dao.TemplateProvider;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import javax.inject.Inject;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -13,25 +13,35 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
-@WebServlet("/gmail-login")
-public class LoginViewServlet extends HttpServlet {
+@WebServlet("/contact")
+public class ContactUsServlet extends HttpServlet {
 
-    Logger logger = LoggerFactory.getLogger(LoginViewServlet.class);
+    private static Logger logger = LoggerFactory.getLogger(ContactUsServlet.class);
 
     @Inject
     private TemplateProvider templateProvider;
 
+    @Inject
+    SessionBean sessionBean;
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        Template template = templateProvider.getTemplate(getServletContext(), "login.ftlh");
+        Template template = templateProvider.getTemplate(getServletContext(), "contactUs.ftlh");
+        Map<String, Object> model = new HashMap<>();
+
+        if(sessionBean.getLogged()){
+            model.put("logged", "yes");
+        }
 
         try {
-            template.process(new Object(), resp.getWriter());
+            template.process(model, resp.getWriter());
         } catch (TemplateException e) {
             e.printStackTrace();
-            logger.warn("Can't load template");
+            logger.warn("View contacts cannot be loaded template!");
         }
     }
 }
