+++ YummyTimeWeb/src/main/java/servlets/menuServlets/StatisticsServlet.java
@@ -1,11 +1,12 @@
-package servletsDoGET;
+package servlets.menuServlets;
 
-import dao.SessionBean;
-import dao.TemplateProvider;
+import dao.*;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import servlets.recipeOperationsServlets.ViewRecipeServlet;
+
 import javax.inject.Inject;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,15 +15,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
-@WebServlet("/contact")
-public class ContactUsServlet extends HttpServlet {
+@WebServlet("/statistics")
+public class StatisticsServlet extends HttpServlet {
 
-    private static Logger logger = LoggerFactory.getLogger(ContactUsServlet.class);
+    private static Logger logger = LoggerFactory.getLogger(ViewRecipeServlet.class);
 
     @Inject
     private TemplateProvider templateProvider;
+
+    @Inject
+    private RecipeDao recipeDao;
 
     @Inject
     SessionBean sessionBean;
@@ -30,8 +35,19 @@ public class ContactUsServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        Template template = templateProvider.getTemplate(getServletContext(), "contactUs.ftlh");
+        Template template = templateProvider.getTemplate(getServletContext(), "statisticsWeb.ftlh");
         Map<String, Object> model = new HashMap<>();
+
+        List<Statistic> categoryStatistic = recipeDao.categoryStatistics();
+        List<Statistic> recipeStatistic = recipeDao.statisticRecipe();
+
+        if (!categoryStatistic.isEmpty()) {
+            model.put("categoryStatistic", categoryStatistic);
+        }
+
+        if (!recipeStatistic.isEmpty()) {
+            model.put("recipeStatistic", recipeStatistic);
+        }
 
         if(sessionBean.getLogged()){
             model.put("logged", "yes");
@@ -41,7 +57,7 @@ public class ContactUsServlet extends HttpServlet {
             template.process(model, resp.getWriter());
         } catch (TemplateException e) {
             e.printStackTrace();
-            logger.warn("View contacts cannot be loaded template!");
+            logger.warn("View statistics cannot be loaded template!");
         }
     }
 }
