+++ YummyTimeWeb/src/main/java/servlets/menuServlets/WelcomeServlet.java
@@ -1,4 +1,4 @@
-package servletsDoGET;
+package servlets.menuServlets;
 
 import com.infoshareacademy.jjdd4.wildhogs.data.Category;
 import dao.BlockRecipe;
