+++ YummyTimeWeb/src/main/java/servlets/recipeOperationsServlets/RecipeToFavoriteServlet.java
@@ -1,6 +1,6 @@
-package servletsDoPOST;
+package servlets.recipeOperationsServlets;
 
-import dao.ShoppingListOfUserDao;
+import dao.RecipeChangeDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,33 +12,33 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/to-shoppingList")
-public class RecipeToShoppingList extends HttpServlet {
+@WebServlet("/to-favorite")
+public class RecipeToFavoriteServlet extends HttpServlet {
 
     @Inject
-    private ShoppingListOfUserDao shoppingListOfUserDao;
+    private RecipeChangeDao recipeChangeDao;
 
     private static Logger logger = LoggerFactory.getLogger(RecipeToFavoriteServlet.class);
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        logger.info("Reading 'id' parameter.");
         String recipeIdParam = req.getParameter("id");
+        logger.info("Reading 'id' parameter.");
 
         if (recipeIdParam == null || recipeIdParam.isEmpty()) {
             logger.info("'id' parameter is invalid.");
             resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             return;
         }
-        Long recipeId = Long.valueOf(recipeIdParam);
+        Long recipeId =  Long.valueOf(recipeIdParam);
 
-        Boolean recipeAdd = shoppingListOfUserDao.addRecipeToShoppingList(recipeId);
+        Boolean recipeAdd = recipeChangeDao.addRecipeToFavorites(recipeId);
 
         String path;
-        if (recipeAdd) {
-            logger.info("Recipe added to shopping list.");
-            path = "/view-recipe?id=" + recipeIdParam + "&shoppingList=yes";
+        if(recipeAdd) {
+            path = "/view-recipe?id=" + recipeIdParam + "&favorite=yes";
+            logger.info("Recipe added to favorites.");
         } else {
             path = "/view-recipe?id=" + recipeIdParam;
         }
