+++ YummyTimeWeb/src/main/java/servlets/recipeOperationsServlets/RecipeToShoppingList.java
@@ -1,6 +1,6 @@
-package servletsDoPOST;
+package servlets.recipeOperationsServlets;
 
-import dao.RecipeChangeDao;
+import dao.ShoppingListOfUserDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,33 +12,33 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/to-favorite")
-public class RecipeToFavoriteServlet extends HttpServlet {
+@WebServlet("/to-shoppingList")
+public class RecipeToShoppingList extends HttpServlet {
 
     @Inject
-    private RecipeChangeDao recipeChangeDao;
+    private ShoppingListOfUserDao shoppingListOfUserDao;
 
     private static Logger logger = LoggerFactory.getLogger(RecipeToFavoriteServlet.class);
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        String recipeIdParam = req.getParameter("id");
         logger.info("Reading 'id' parameter.");
+        String recipeIdParam = req.getParameter("id");
 
         if (recipeIdParam == null || recipeIdParam.isEmpty()) {
             logger.info("'id' parameter is invalid.");
             resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             return;
         }
-        Long recipeId =  Long.valueOf(recipeIdParam);
+        Long recipeId = Long.valueOf(recipeIdParam);
 
-        Boolean recipeAdd = recipeChangeDao.addRecipeToFavorites(recipeId);
+        Boolean recipeAdd = shoppingListOfUserDao.addRecipeToShoppingList(recipeId);
 
         String path;
-        if(recipeAdd) {
-            path = "/view-recipe?id=" + recipeIdParam + "&favorite=yes";
-            logger.info("Recipe added to favorites.");
+        if (recipeAdd) {
+            logger.info("Recipe added to shopping list.");
+            path = "/view-recipe?id=" + recipeIdParam + "&shoppingList=yes";
         } else {
             path = "/view-recipe?id=" + recipeIdParam;
         }
